refactoring code and test passed with POSTMAN
