+++ app.js
@@ -4,7 +4,8 @@ const favicon = require("serve-favicon");
 const bodyParser = require("body-parser");
 const config = require("./src/db/config.json");
 const db = require("./src/db/db_config");
-
+const adminRoutes = require("./src/routes/admin");
+const appRoutes = require("./src/routes/appRoutes");
 //import create table
 const {
   createImageTableIfNotExists,
@@ -32,7 +33,9 @@ app
   //middleware pour ameliorer la lisibilité des reponses des requetes
   .use(morgan("dev"))
   //parse body en JSON
-  .use(bodyParser.json());
+  .use(bodyParser.json())
+  .use("/admin", adminRoutes)
+  .use(appRoutes);
 
 //mise en place des tables SQL
 
@@ -48,51 +51,6 @@ createOpening_hoursTableIfNotExists();
 
 require("./src/controllers/auth.js")(app, db);
 
-/*----------------------------------------------------------------------------
---------------------- CRUD OPERATION SECOND HAND CAR--------------------------
------------------------------------------------------------------------------*/
-require("./src/routes/Second-hand-car-routes/createCar.js")(app, db);
-require("./src/routes/Second-hand-car-routes/getAllCars.js")(app, db);
-require("./src/routes/Second-hand-car-routes/getCarByID.js")(app, db);
-require("./src/routes/Second-hand-car-routes/deleteCar.js")(app, db);
-require("./src/routes/Second-hand-car-routes/updateCar.js")(app, db);
-/* ---------------------------------------------------------------------------
--------------------------CRUD OPERATION HOURS REQUEST-------------------------
------------------------------------------------------------------------------- */
-
-require("./src/routes/Hours-routes/getHours.js")(app, db);
-require("./src/routes/Hours-routes/updateHours.js")(app, db);
-
-/* ---------------------------------------------------------------------------
---------------------------CRUD OPERATION REVIEW REQUEST-----------------------
------------------------------------------------------------------------------- */
-require("./src/routes/Review-routes/getAllReview.js")(app, db);
-require("./src/routes/Review-routes/createReview.js")(app, db);
-require("./src/routes/Review-routes/getValidedReviews.js")(app, db);
-require("./src/routes/Review-routes/getPendingReviews.js")(app, db);
-require("./src/routes/Review-routes/approvedPendingReview.js")(app, db);
-require("./src/routes/Review-routes/deletedReview.js")(app, db);
-
-/* ---------------------------------------------------------------------------
---------------------------CRUD OPERATION CONTACT REQUEST----------------------
------------------------------------------------------------------------------- */
-
-require("./src/routes/contact-routes/createContact.js")(app, db);
-require("./src/routes/contact-routes/getAllContact.js")(app, db);
-require("./src/routes/contact-routes/deleteContact.js")(app, db);
-
-/* ---------------------------------------------------------------------------
----------------------------CRUD OPERATION HOME PAGE REQUEST-------------------
------------------------------------------------------------------------------- */
-
-require("./src/routes/HomePage-routes/getAllContaintHome.js")(app, db);
-require("./src/routes/HomePage-routes/updatedContaintHomePageByID.js")(app, db);
-
-/* ---------------------------------------------------------------------------
------------------------------------SECTION PAGE REQUEST------------------------------
------------------------------------------------------------------------------- */
-require("./src/routes/section-routes/getAllSections.js")(app, db);
-require("./src/routes/section-routes/updateSection.js")(app, db);
 
 app.listen(config.port, () =>
   console.log(`node started to port ${config.port}`)
